requirements.txt
amqp==2.5.2
astroid==2.3.3
billiard==3.6.3.0
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.8.0
django-heroku==0.3.1
gunicorn==20.0.0
idna==2.9
importlib-metadata==1.5.1
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.8
vine==1.3.0
whitenoise==4.1.4
wrapt==1.11.2
zipp==3.1.0
